% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make_cleanup}
\alias{make_cleanup}
\title{Cleanup Rare Phenotypes}
\usage{
make_cleanup(eps)
}
\arguments{
\item{eps}{Abundance below which species are considered extinct.}
}
\description{
Cleanup phenotypes that have dropped below a threshold abundance
}
\author{
Rich FitzJohn
}

