\name{make_dieckmann_1999}
\alias{make_dieckmann_1999}
\title{Dieckmann & Doebeli 1999}
\usage{
make_dieckmann_1999(r = 1, K_0 = 500, s2_C = 0.16, s2_K = 1)
}
\arguments{
  \item{r}{Intrinsic birth rate}

  \item{K_0}{Scaling factor giving maximum population
  density}

  \item{s2_C}{Width of the competition kernel}

  \item{s2_K}{Width of the resource distribution kernel}
}
\value{
TBD...
}
\description{
Model implementing "On the origin of species by sympatric
speciation" by Dieckmann & Doebeli, Nature 1999.
}
\details{
Still a test case only; much to come.
}
\author{
Rich FitzJohn
}

