\name{make_huisman_2001}
\alias{make_huisman_2001}
\title{Huisman & Weissing 2001}
\usage{
make_huisman_2001(m = 0.25, D = 0.25, S = c(1, 1))
}
\arguments{
  \item{d}{See paper}

  \item{D}{See paper}

  \item{S}{See paper}
}
\description{
Model implementing "Huisman, J., and Weissing, F.J. (2001).
Biological conditions for oscillations and chaos generated
by multispecies competition. Ecology 82, 2682–2695..
}
\details{
Test case...
}
\author{
Georges Kunstler and Rich FitzJohn
}

