\name{make_rstar}
\alias{make_rstar}
\alias{make_rstar_identity}
\alias{make_rstar_mat_constant}
\alias{rstar_mat_1}
\alias{rstar_mat_2}
\alias{rstar_mat_2_tradeoff}
\alias{rstar_matrices}
\alias{rstar_matrices_fixed}
\title{R* model}
\usage{
make_rstar(matrices, r = 1, m = 0.25, D = 0.25, S = 1)

make_rstar_identity(k)

rstar_mat_1(x)

rstar_mat_2(x)

rstar_mat_2_tradeoff(x)

make_rstar_mat_constant(M)

rstar_matrices_fixed(K, C)

rstar_matrices(K, C)
}
\arguments{
  \item{r}{Intrinsic growth rate}

  \item{m}{Mortality rate}

  \item{D}{Resource turnover rate}

  \item{S}{Supply rate}

  \item{matrices}{Matrices generated by
  \code{rstar_matrices}}

  \item{k}{number of resources}

  \item{x}{matrix of traits (for the \code{rstar_mat_}
  functions).}

  \item{M}{coumn matrix with K or C matrix that will be
  shared by all species.}

  \item{K}{Matrix of Monod constants}

  \item{C}{Matrix of consumption constants}
}
\description{
Model implementing the "R*" model, most closely based on
that presented by "Huisman, J., and Weissing, F.J. (2001).
Biological conditions for oscillations and chaos generated
by multispecies competition. Ecology 82, 2682-2695."
}
\details{
Warning: there are some dependent parameters here, and they
won't be updated until I sort out some sort of parameter
hook.  Or I'll just remove the ability to set parameters
from the parmeters object, which reduces the complexity a
bit.

So, what we need to have is a couple of "axes" for
parameters to vary over.  But the actual parameters
underlying things are a crazy matrix of values.
}
\author{
Georges Kunstler and Rich FitzJohn
}

