% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{rstar-class}
\alias{make_rstar_identity}
\alias{make_rstar_mat_constant}
\alias{rstar}
\alias{rstar-class}
\alias{rstar_mat_1}
\alias{rstar_mat_2}
\alias{rstar_mat_2_tradeoff}
\alias{rstar_matrices}
\alias{rstar_matrices_fixed}
\title{R* model

Model implementing the "R*" model, most closely based on that
presented by "Huisman, J., and Weissing, F.J. (2001).  Biological
conditions for oscillations and chaos generated by multispecies
competition. Ecology 82, 2682-2695."}
\usage{
make_rstar_identity(k)

rstar_mat_1(x)

rstar_mat_2(x)

rstar_mat_2_tradeoff(x)

make_rstar_mat_constant(M)

rstar_matrices_fixed(K, C)

rstar_matrices(K, C)
}
\arguments{
\item{k}{number of resources}

\item{x}{matrix of traits (for the \code{rstar_mat_} functions).}

\item{M}{coumn matrix with K or C matrix that will be shared by
all species.}

\item{K}{Matrix of Monod constants}

\item{C}{Matrix of consumption constants}
}
\description{
R* model

Model implementing the "R*" model, most closely based on that
presented by "Huisman, J., and Weissing, F.J. (2001).  Biological
conditions for oscillations and chaos generated by multispecies
competition. Ecology 82, 2682-2695."
}
\section{Methods}{

\describe{
\item{\code{Rstar(x)}}{Compute equilibrium level of each resource (i.e. level it will
reach when limiting.  This is the resource level where the growth rate
equals the mortality rate.}

\item{\code{carrying_capacity(x)}}{Compute the carrying capacity -- the population size that a species
with trait x will reach when alone.}

\item{\code{dNdt(x, N, R)}}{Change in density with respect to time (total, not per capita)}

\item{\code{dRdt(x, N, R)}}{Change in resources with respect to time}

\item{\code{equilibrium(sys, ...)}}{Given traits and initial densities as a system (sys) compute equilibirum numerically.}

\item{\code{equilibrium_R(sys, ...)}}{Given traits 'x' and initial densities 'y', compute the
equilibrium.  Return the equilibrium resource availability at the
same time.}

\item{\code{p(x, R)}}{Specific growth rate at a particular resource level R}

\item{\code{pinv(x, g)}}{Resource level that would be required for particular growth rate g}

\item{\code{single_equilibrium(x)}}{Given species traits \code{x}, compute equilibrium system.  If multiple species are given, the output will be multiple separate equilibira, rather than an actual community.}

\item{\code{single_equilibrium_R(sys, force.numerical = FALSE)}}{Solve for the equilibrium level of a resource, given a vector of
densities (and species traits).  Do do this, we solve dRdt == 0:
    $D (S - R) - c N p(R) == 0$
    $D * (S - R) - c N r R / (K + R) == 0$
 As a quadratic with respect to 'R':
    $-D * R^2 + (D (S - K) - c r N) * R + D K S == 0$}
}}

